+++ config.h
@@ -1,7 +1,8 @@
 #pragma once
 
-#include <stdint.h>
+#include <regex.h>
 #include <stdbool.h>
+#include <stdint.h>
 #include <uchar.h>
 
 #include <xkbcommon/xkbcommon.h>
@@ -60,6 +61,7 @@ enum binding_aux_type {
     BINDING_AUX_NONE,
     BINDING_AUX_PIPE,
     BINDING_AUX_TEXT,
+    BINDING_AUX_REGEX,
 };
 
 struct binding_aux {
@@ -73,6 +75,8 @@ struct binding_aux {
             uint8_t *data;
             size_t len;
         } text;
+
+        char *regex_name;
     };
 };
 
@@ -119,6 +123,13 @@ struct env_var {
     char *value;
 };
 typedef tll(struct env_var) env_var_list_t;
+
+struct custom_regex {
+    char *name;
+    char *regex;
+    regex_t preg;
+    struct config_spawn_template launch;
+};
 
 struct config {
     char *term;
@@ -220,11 +231,11 @@ struct config {
             OSC8_UNDERLINE_ALWAYS,
         } osc8_underline;
 
-        char32_t **protocols;
-        char32_t *uri_characters;
-        size_t prot_count;
-        size_t max_prot_len;
+        char *regex;
+        regex_t preg;
     } url;
+
+    tll(struct custom_regex) custom_regexes;
 
     struct {
         uint32_t fg;